backend/requirements.txt
fastapi
uvicorn
websockets
googlemaps
pydantic
python-dotenv
polyline
geopy
